requirements.txt
farm-haystack[inference]
datasets>=2.6.1
python-dotenv
chainlit
farm-haystack[weaviate]